refactor: move clients package
