+++ internal/whisper/clients/mock/client.go
