+++ internal/whisper/clients/passthrough/client.go
